requirements.txt
scipy
numpy
scikit-learn
tensorlayer
tensorflow
Pillow
tkFileDialog

